renamed f16 to bf16 for clarity

+++ ppci/arch/atalla/registers.py
@@ -140,7 +140,7 @@ gdb_registers = registers + [PC]
 register_classes_swfp = [
     RegisterClass(
         "reg",
-        [ir.i8, ir.i16, ir.i32, ir.ptr, ir.u8, ir.u16, ir.u32, ir.f16],
+        [ir.i8, ir.i16, ir.i32, ir.ptr, ir.u8, ir.u16, ir.u32, ir.bf16],
         AtallaRegister,
         [
             R9,
